Fixes #1856 - Move TestUtils method related to security headers to SecurityUtil
