+++ catalog-rest-service/src/main/java/org/openmetadata/catalog/resources/charts/ChartResource.java
@@ -31,7 +31,6 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 import javax.json.JsonPatch;
@@ -95,7 +94,6 @@ public class ChartResource {
 
   @Inject
   public ChartResource(CollectionDAO dao, CatalogAuthorizer authorizer) {
-    Objects.requireNonNull(dao, "CollectionDAO must not be null");
     this.dao = new ChartRepository(dao);
     this.authorizer = authorizer;
   }
