+++ catalog-rest-service/src/main/resources/json/schema/type/changeEvent.json
@@ -33,7 +33,8 @@
           }
         }
       },
-      "required": ["eventType"]
+      "required": ["eventType"],
+      "additionalProperties": false
     }
   },
   "properties" : {
